Endpoint to get registration mandatory fields
