+++ events/api_translated.py
@@ -1,0 +1,108 @@
+from django.conf import settings
+from modeltranslation.translator import NotRegistered, translator
+from rest_framework import serializers
+
+from events import utils
+
+
+class TranslatedModelSerializer(serializers.ModelSerializer):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        model = self.Meta.model
+        try:
+            trans_opts = translator.get_options_for_model(model)
+        except NotRegistered:
+            self.translated_fields = []
+            return
+
+        self.translated_fields = trans_opts.fields.keys()
+        lang_codes = utils.get_fixed_lang_codes()
+        # Remove the pre-existing data in the bundle.
+        for field_name in self.translated_fields:
+            for lang in lang_codes:
+                key = "%s_%s" % (field_name, lang)
+                if key in self.fields:
+                    del self.fields[key]
+            del self.fields[field_name]
+
+    def to_representation(self, obj):
+        ret = super().to_representation(obj)
+        if obj is None:
+            return ret
+        return self.translated_fields_to_representation(obj, ret)
+
+    def to_internal_value(self, data):
+        """
+        Convert complex translated json objects to flat format.
+        E.g. json structure containing `name` key like this:
+        {
+            "name": {
+                "fi": "musiikkiklubit",
+                "sv": "musikklubbar",
+                "en": "music clubs"
+            },
+            ...
+        }
+        Transforms this:
+        {
+            "name": "musiikkiklubit",
+            "name_fi": "musiikkiklubit",
+            "name_sv": "musikklubbar",
+            "name_en": "music clubs"
+            ...
+        }
+        :param data:
+        :return:
+        """
+
+        extra_fields = {}  # will contain the transformation result
+        for field_name in self.translated_fields:
+            obj = data.get(field_name, None)  # { "fi": "musiikkiklubit", "sv": ... }
+            if not obj:
+                continue
+            if not isinstance(obj, dict):
+                raise serializers.ValidationError(
+                    {
+                        field_name: "This field is a translated field. Instead of a string,"
+                        " you must supply an object with strings corresponding"
+                        " to desired language ids."
+                    }
+                )
+            for language in (
+                lang for lang in utils.get_fixed_lang_codes() if lang in obj
+            ):
+                value = obj[language]  # "musiikkiklubit"
+                if language == settings.LANGUAGES[0][0]:  # default language
+                    extra_fields[field_name] = value  # { "name": "musiikkiklubit" }
+                extra_fields[
+                    "{}_{}".format(field_name, language)
+                ] = value  # { "name_fi": "musiikkiklubit" }
+            del data[field_name]  # delete original translated fields
+
+        # handle other than translated fields
+        data = super().to_internal_value(data)
+
+        # add translated fields to the final result
+        data.update(extra_fields)
+
+        return data
+
+    def translated_fields_to_representation(self, obj, ret):
+        for field_name in self.translated_fields:
+            d = {}
+            for lang in utils.get_fixed_lang_codes():
+                key = "%s_%s" % (field_name, lang)
+                val = getattr(obj, key, None)
+                if val is None:
+                    continue
+                d[lang] = val
+
+            # If no text provided, leave the field as null
+            for _key, val in d.items():
+                if val is not None:
+                    break
+            else:
+                d = None
+            ret[field_name] = d
+
+        return ret
