+++ helevents/templates/rest_framework/api.html
@@ -33,6 +33,8 @@
     {% include "rest_framework/language_list.html" %}
 {% elif name == "Organization List" %}
     {% include "rest_framework/organization_list.html" %}
+{% elif name == "Mandatory Field List" %}
+    {% include "rest_framework/registration_mandatory_field_list.html" %}
 {% else %}
 {% endif %}
 {% endblock %}
