+++ helevents/templates/rest_framework/registration_mandatory_field_list.html
@@ -1,0 +1,17 @@
+<div class="panel panel-default">
+    <div class="panel-body">
+        <h2 id="using-registration-mandatory-field-endpoint">Using the registration mandatory field endpoint</h2>
+        <p>Here, the registration mandatory field options are listed. Registration mandatory fields are used to 
+            define mandatory fields per registration and defined fields will be validated during the enrolment.</p>
+        
+        <h3 id="mandatory-field-type">Mandatory field type</h3>
+        <p>To find out mandatory fields that has a specific type, use the query parameter
+            <code>type</code>, separating values by commas if you wish to query for several type.</p>
+        <p>Example:</p>
+        <pre><code>registration-mandatory-field/?type=contact</code></pre>
+        <p><a href="?type=person" title="json">See the result</a></p>
+    </div>
+</div>
+
+
+
